Support kDNSServiceFlagsMoreComing
If we want to use this library as a regular resolver we need to know when
the stream of results is "finished", so we can return the results to the
application.

This patch maps kDNSServiceFlagsMoreComing to a new flag `more`

https://developer.apple.com/documentation/dnssd/1823436-anonymous/kdnsserviceflagsmorecoming

A resolver can wait until the `more` flag is `false` and return the results
as a batch.

Signed-off-by: David Scott <dave.scott@docker.com>